+++ crates/top/re_sdk/src/log_sink.rs
@@ -593,54 +593,3 @@ impl LogSink for GrpcSink {
         self
     }
 }
-
-// ----------------------------------------------------------------------------
-//
-// TODO:
-// * Move to own file.
-// * Better names.
-
-pub trait PipelineTransform: Send + Sync + 'static {
-    fn apply(&self, msg: LogMsg) -> Vec<LogMsg>;
-
-    fn to_sink<S: LogSink>(self, sink: S) -> Pipeline<S, Self>
-    where
-        Self: Sized,
-    {
-        Pipeline {
-            sink,
-            transform: self,
-        }
-    }
-}
-
-/// A sink which can transform a `LogMsg` and forward the result to an underlying backing `LogSink`.
-pub struct Pipeline<S: LogSink, T: PipelineTransform> {
-    sink: S,
-    transform: T,
-}
-
-impl<S: LogSink, T: PipelineTransform> Pipeline<S, T> {
-    /// Create a new `TransformSink` with the given transform function.
-    #[inline]
-    pub fn new<F>(sink: S, transform: T) -> Self {
-        Self { sink, transform }
-    }
-}
-
-impl<S: LogSink, T: PipelineTransform> LogSink for Pipeline<S, T> {
-    fn send(&self, msg: re_log_types::LogMsg) {
-        self.sink.send_all(self.transform.apply(msg))
-    }
-
-    fn flush_blocking(
-        &self,
-        timeout: std::time::Duration,
-    ) -> Result<(), crate::sink::SinkFlushError> {
-        self.sink.flush_blocking(timeout)
-    }
-
-    fn as_any(&self) -> &dyn std::any::Any {
-        self
-    }
-}
