+++ examples/rust/lenses/README.md
@@ -2,7 +2,9 @@
 title = "Transform recording stream example"
 -->
 
+<!-- TODO: -->
+
 Demonstrates how to transform log messages before forwarding them to the sink the SDK.
 ```bash
-cargo run -p transform
+cargo run -p lenses
 ```
